+++ src/main/java/net/orandja/strawberry/blockentity/TeleporterBlockEntity.java
@@ -21,6 +21,8 @@ import net.orandja.chocoflavor.ChocoFlavor;
 import net.orandja.chocoflavor.utils.GlobalUtils;
 import net.orandja.strawberry.StrawberryCustomBlocks;
 import net.orandja.strawberry.intf.StrawberryBlockEntity;
+import net.orandja.strawberry.intf.StrawberryMarkerEntity;
+import net.orandja.strawberry.intf.StrawberryPlayer;
 import net.orandja.strawberry.item.TeleportingEssence;
 
 import java.util.ArrayList;